python/fastapi-celery-postgres/pyproject.toml
[project]
name = "fastapi-celery-postgres"
version = "0.1.0"
description = "FastAPI + Celery + PostgreSQL OpenTelemetry example"
authors = [
    {name = "Irfan Shah",email = "irfns@protonmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"


dependencies = [
    "uvicorn[standard] (>=0.41.0,<1.0.0)",
    "fastapi (>=0.133.1,<1.0.0)",
    "celery (>=5.6.0,<6.0.0)",
    "redis (>=7.1.0,<8.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "sqlalchemy (>=2.0.47,<3.0.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "opentelemetry-distro (>=0.60b1)",
    "opentelemetry-exporter-otlp (>=1.39.1,<2.0.0)",
    "opentelemetry-instrumentation-celery (>=0.60b1)",
    "opentelemetry-instrumentation-fastapi (>=0.60b1)",
    "opentelemetry-instrumentation-sqlalchemy (>=0.60b1)",
    "opentelemetry-instrumentation-redis (>=0.60b1)",
    "opentelemetry-instrumentation-logging (>=0.60b1)"
]


[project.optional-dependencies]
dev = [
    "pytest (>=9.0,<10.0)",
    "pytest-cov (>=7.0,<8.0)",
    "httpx (>=0.28,<1.0)",
    "ruff (>=0.9)",
    "mypy (>=1.14)",
]

[tool.poetry]
packages = [{include = "app"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py313"
line-length = 100

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # function call in default argument (FastAPI Depends)
    "ARG001", # unused function argument (FastAPI/Celery handlers)
    "ARG002", # unused method argument (pytest fixtures)
]

[tool.ruff.lint.isort]
known-first-party = ["app"]

# MyPy - type checking (non-strict: legacy SQLAlchemy + Celery lack stubs)
[tool.mypy]
python_version = "3.13"
warn_unused_configs = true
warn_redundant_casts = true
check_untyped_defs = true
show_error_codes = true
pretty = true

[[tool.mypy.overrides]]
module = [
    "celery.*",
    "redis.*",
    "opentelemetry.*",
    "sqlalchemy.*",
    "environ.*",
]
ignore_missing_imports = true

# Legacy SQLAlchemy Column() style and Celery tasks
[[tool.mypy.overrides]]
module = ["app.telemetry.*", "app.config", "app.tasks"]
ignore_errors = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"